ENH: Adding ITKHighDimensionalOptimizers and ITKHighDimensionalMetrics
For creating a gerrit patch, the new modules will be copied from this
branch into ITK master, and the new modules added to ITKGroups list.

This is the first step in moving the PICSL RegistrationRefactoring new
registration framework into ITK master.

HighDimensionalMetrics module was also added to include
ObjectToObjectMetric, which is used in ObjectToObjectOptimizer.

Three threading-helper classes were also added: ObjectToData, ImageToData
and Array1DToData, because of their use in the new Optimziers and metrics.
If they're kept, they should be moved to a different module.

Change-Id: I06dda4c9203e7921778c969fa6d3db452729afec